Added another warning flag, -Wshorten-64-to-32 and fixed more compiler warnings.

+++ NSStringAdditions.m
@@ -24,7 +24,7 @@
 
 - (unsigned int)unsignedIntValue
 {
-	return strtoul([self UTF8String], NULL, 10);
+	return (unsigned int)strtoul([self UTF8String], NULL, 10);
 }
 
 - (unsigned long long)unsignedLongLongValue
